Made the makefile a bit more dynamic
I was getting annoyed at having to change the first line of the makefile
every time I added it to a new project. So, I set up the file to read
from an environment variable. Now, I can have the makefile in one spot,
symlink it to my project, and add a .makedoc file that will set the
template variable for me.

I updated the README file with some information regarding this.