Restored tldiv variable
Had to do a --chapters workaround briefly because I didn't have an up to
date version of Pandoc (and was way too lazy to upgrade until now). The
changes since then ended up clobbering the tldiv, so I've add it back
in, since that is a better way to handle compiling a document.