Added some extra scripts
Added an invocation to an after.sh script to allow the user to add some
post-build tasks (such as copying the compiled document to Dropbox).
Started a setup script and a todo parsing script that will create a JSON
file.

There is a lot more to do (some of these TODOs are in the comments of
the files, some are not).