Edits to IDD LaTeX template
Droped the test cases sections. It's better to just write the test cases
than it is to write about the test cases.

Added a LaTeX todo list package, so tasks can be added in-line while
writing (and then culled to the beginning of the document).

Added title page image and embedded the Barnard's Star in the templates
module.