Added TWI Hashing
When MakeDoc parses todos out of an IDD into Taskwarrior, it will
generate an MD5 Hash of the task the same way TWInput does. If
it finds that hash in the pending tasks, it won't insert the task again.

This uses the same UDA that TWInput uses for the hash. So, this feature
should be compatbile with TWInput.